Bug 1892913 - patch 1 - Update cairo source to 1.18.0 r=gfx-reviewers,lsalzman
Wholesale replacement of files in gfx/cairo/cairo with contents of the 1.18.0 release tarball,
omitting subdirectories that are not part of the core library build.

(This will not build as-is; the following patches will fix things up for Gecko integration.)

Differential Revision: https://phabricator.services.mozilla.com/D209407

+++ gfx/cairo/cairo/src/cairo-hash.c
@@ -340,7 +340,7 @@ _cairo_hash_table_lookup (cairo_hash_table_t *hash_table,
 {
     cairo_hash_entry_t *entry;
     unsigned long table_size, i, idx, step;
-    unsigned long hash = key->hash;
+    uintptr_t hash = key->hash;
 
     entry = hash_table->cache[hash & 31];
     if (entry && entry->hash == hash && hash_table->keys_equal (key, entry))
