+++ gfx/cairo/cairo/src/cairo-paginated-private.h
@@ -178,7 +178,7 @@ _cairo_surface_is_paginated (cairo_surface_t *surface);
 
 cairo_private cairo_status_t
 _cairo_paginated_surface_set_size (cairo_surface_t 	*surface,
-				   int			 width,
-				   int			 height);
+				   double		 width,
+				   double		 height);
 
 #endif /* CAIRO_PAGINATED_H */
